refactor: move common.js into transformers
